requirements.txt
click
pydicom
pynetdicom
tqdm
rich
PyYAML
